오더 리스트

+++ order-service/src/main/java/com/green/eats/order/application/model/OrderPostItemReq.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class OrderPostItemReq {
     private Long menuId;
     private Integer quantity;
-    private Long price; // 단가
+    private Integer price; // 단가
 }
