+++ order-service/src/main/java/com/green/eats/order/application/model/OrderPostReq.java
@@ -16,5 +16,5 @@ public class OrderPostReq {
     private List<OrderPostItemReq> items;
 
     @Positive(message = "0 이상이어야 합니다.")
-    private Long totalAmount;
+    private Integer totalAmount;
 }
